BackEnd/music_history/pyproject.toml
[tool.poetry]
name = "mht"
version = "0.1.0"
description = ""
authors = ["Pablo <godoypablojuan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
